[IMP] estate: add estate contract feature and fix style issues
