+++ awesome_dashboard/static/src/dashboard.scss
@@ -1,0 +1,3 @@
+.o_dashboard {
+    background-color: #f0f0f0;
+}
