+++ estate_contract/models/__init__.py
@@ -1,0 +1,2 @@
+from . import estate_contract
+from . import estate_property
