+++ estate_contract/security/ir.model.access.csv
@@ -1,0 +1,2 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_estate_contract_user,estate.contract.user,model_estate_contract,base.group_user,1,1,1,1
