+++ estate_contract/views/estate_contract_views.xml
