+++ estate_contract/views/estate_property_views.xml
@@ -6,15 +6,15 @@
         <field name="model">estate.property</field>
         <field name="inherit_id" ref="estate.estate_property_form_view"/>
         <field name="arch" type="xml">
-
             <xpath expr="//div[@name='button_box']" position="inside">
-                <button name="action_open_event" type="object" class="oe_stat_button" icon="fa-calendar">
+
+                <button name="action_open_contract" type="object" class="oe_stat_button" icon="fa-file" invisible="state != 'offer_accepted'">
                     <div class="o_stat_info">
-                        <span class="o_stat_text">Event</span>
+                        <span class="o_stat_text">Contract</span>
                     </div>
                 </button>
+
             </xpath>
-            
         </field>
     </record>
 </odoo>
