+++ estate_event/models/__init__.py
@@ -1,3 +1,2 @@
 from . import estate_property
 from . import estate_property_offer
-from . import estate_contract
