+++ estate_event/models/estate_property_offer.py
@@ -1,6 +1,7 @@
 from odoo import models, api, fields
 from odoo.exceptions import UserError
 
+
 class EstatePropertyOffer(models.Model):
     _inherit = 'estate.property.offer'
 
@@ -34,7 +35,7 @@ class EstatePropertyOffer(models.Model):
                 raise UserError("Buyer must attend the event before making an offer.")
 
         return super().create(vals_list)
-    
+
     def _compute_allowed_partner_ids(self):
         for record in self:
             if not record.property_id.event_id:
